+++ src/Plc.Test/ItemTests/DynamicPlcItemTest.cs
@@ -11,56 +11,101 @@ namespace Phoenix.Data.Plc.Test.ItemTests
 	public sealed class DynamicPlcItemTest
 	{
 		/// <summary>
-		/// Checks if a custom <see cref="IDynamicPlcItem.LengthLimit"/> is respected when reading such an item.
+		/// Checks if a custom <see cref="IDynamicPlcItem.LengthFactor"/> is applied when reading an <see cref="IDynamicPlcItem"/>.
 		/// </summary>
 		[Test]
-		public void Check_If_Limit_Is_Respected_When_Reading()
+		[TestCase((ushort) 65535)]
+		public void Check_If_Factor_Is_Applied_When_Reading(ushort value)
 		{
 			// Arrange
-			var limit = (uint) 2;
-			var numericItem = new BytePlcItem(0, 0);
-			Func<string, IPlcItem<byte[]>> flexiblePlcItemFactory = name => new BytesPlcItem(0, 1, true, new byte[4], name);
-			// ReSharper disable once ObjectCreationAsStatement → It is necessary to wrap both the numeric and the flexible item to check if the limit supplied to the constructor of the dynamic item is respected.
-			new DynamicPlcItem<byte[]>(numericItem, flexiblePlcItemFactory, limit);
+			var factor = (byte) 2;
+			var targetDataLength = (uint) value * factor;
+			var numericItem = new UInt16PlcItem(0, 0);
+			Func<string, IPlcItem<byte[]>> flexiblePlcItemFactory = name => new BytesPlcItem(0, 1, true, new byte[0], name);
+			var dynamicPlcItem = new DynamicPlcItem<byte[]>(numericItem, flexiblePlcItemFactory, factor, null);
 
 			// Act
-			numericItem.Value = 3;
+			numericItem.Value = value;
 
 			// Assert
-			Assert.AreEqual(numericItem.Value, limit);
-			
-			// Act
-			numericItem.Value = 1;
-
-			// Assert
-			Assert.AreEqual(numericItem.Value, 1);
+			Assert.That(numericItem.Value, Is.EqualTo(value));
+			Assert.That(dynamicPlcItem.FlexiblePlcItem.Value, Has.Length.EqualTo(targetDataLength));
 		}
 
 		/// <summary>
-		/// Checks if a custom <see cref="IDynamicPlcItem.LengthLimit"/> is respected when writing such an item.
+		/// Checks if a custom <see cref="IDynamicPlcItem.LengthFactor"/> is applied when writing an <see cref="IDynamicPlcItem"/>.
 		/// </summary>
 		[Test]
-		public void Check_If_Limit_Is_Respected_When_Writing()
+		[TestCase(new byte[]{0,0,0,0,0,0,0,0,0,0})]
+		public void Check_If_Factor_Is_Applied_When_Writing(byte[] value)
+		{
+			// Arrange
+			var factor = (byte) 2;
+			var targetLength = value.Length / factor;
+			var numericItem = new UInt16PlcItem(0, 0);
+			Func<string, IPlcItem<byte[]>> flexiblePlcItemFactory = name => new BytesPlcItem(0, 1, true, new byte[0], name);
+			var dynamicPlcItem = new DynamicPlcItem<byte[]>(numericItem, flexiblePlcItemFactory, factor, null);
+
+			// Act
+			dynamicPlcItem.FlexiblePlcItem.Value = value;
+
+			// Assert
+			Assert.That(numericItem.Value, Is.EqualTo(targetLength));
+			Assert.That(dynamicPlcItem.FlexiblePlcItem.Value, Has.Length.EqualTo(value.Length));
+		}
+
+		/// <summary>
+		/// Checks if a custom <see cref="IDynamicPlcItem.LengthLimit"/> is respected when reading an <see cref="IDynamicPlcItem"/>.
+		/// </summary>
+		[Test]
+		[TestCase(DataConverter.Endianness.LittleEndian, (ushort) 1000, (ushort) 100)] // Value is larger than limit → Limit is applied.
+		[TestCase(DataConverter.Endianness.LittleEndian, (ushort) 50, (ushort) 50)] // Value is smaller than limit → Value will be used.
+		[TestCase(DataConverter.Endianness.BigEndian, (ushort) 1000, (ushort) 100)] // Value is larger than limit → Limit is applied.
+		[TestCase(DataConverter.Endianness.BigEndian, (ushort) 50, (ushort) 50)] // Value is smaller than limit → Value will be used.
+		public void Check_If_Limit_Is_Respected_When_Reading(DataConverter.Endianness endianness, ushort value, ushort target)
+		{
+			// Arrange
+			var limit = (uint) 100;
+			INumericPlcItem numericItem;
+			if (endianness == DataConverter.Endianness.LittleEndian)
+			{
+				numericItem = new UInt16PlcItem(0,0);
+			}
+			else
+			{
+				numericItem = new WordPlcItem(0, 0);
+			}
+			Func<string, IPlcItem<byte[]>> flexiblePlcItemFactory = name => new BytesPlcItem(0, 1, true, new byte[0], name);
+			var dynamicPlcItem = new DynamicPlcItem<byte[]>(numericItem, flexiblePlcItemFactory, 1, limit);
+
+			// Act
+			numericItem.Value = value;
+
+			// Assert
+			Assert.That(numericItem.Value, Is.EqualTo(target));
+			Assert.That(dynamicPlcItem.FlexiblePlcItem.Value, Has.Length.EqualTo(target));
+		}
+
+		/// <summary>
+		/// Checks if a custom <see cref="IDynamicPlcItem.LengthLimit"/> is respected when writing an <see cref="IDynamicPlcItem"/>.
+		/// </summary>
+		[Test]
+		[TestCase(new byte[] { 255, 255, 255, 255 }, 2, new byte[] { 255, 255 })] // This should be limited to only 2 bytes.
+		[TestCase(new byte[] { 0 }, 1, new byte[] { 0 })] // This can be written in total.
+		public void Check_If_Limit_Is_Respected_When_Writing(byte[] value, byte targetLength, byte[] targetData)
 		{
 			// Arrange
 			var limit = (uint) 2;
 			var numericItem = new BytePlcItem(0, 0);
-			Func<string, IPlcItem<byte[]>> flexiblePlcItemFactory = name => new BytesPlcItem(0, 1, true, new byte[4], name);
-			DynamicPlcItem<byte[]> dynamicPlcItem = new DynamicPlcItem<byte[]>(numericItem, flexiblePlcItemFactory, limit);
+			Func<string, IPlcItem<byte[]>> flexiblePlcItemFactory = name => new BytesPlcItem(0, 1, true, new byte[0], name);
+			var dynamicPlcItem = new DynamicPlcItem<byte[]>(numericItem, flexiblePlcItemFactory, 1, limit);
 
 			// Act
-			dynamicPlcItem.FlexiblePlcItem.Value = new [] {byte.MaxValue, byte.MaxValue, byte.MaxValue, byte.MaxValue};
+			dynamicPlcItem.FlexiblePlcItem.Value = value;
 			
 			// Assert
-			Assert.AreEqual(numericItem.Value, limit);
-			Assert.True(new [] {byte.MaxValue, byte.MaxValue}.SequenceEqual(dynamicPlcItem.FlexiblePlcItem.Value));
-
-			// Act
-			dynamicPlcItem.FlexiblePlcItem.Value = new[] { byte.MinValue };
-
-			// Assert
-			Assert.AreEqual(numericItem.Value, 1);
-			Assert.True(new[] { byte.MinValue }.SequenceEqual(dynamicPlcItem.FlexiblePlcItem.Value));
+			Assert.That(numericItem.Value, Is.EqualTo(targetLength));
+			Assert.True(targetData.SequenceEqual(dynamicPlcItem.FlexiblePlcItem.Value));
 		}
 	}
 }