+++ src/Plc/Items/Typed/DynamicUtf8PlcItem.cs
@@ -29,20 +29,20 @@ namespace Phoenix.Data.Plc.Items.Typed
 
 		/// <inheritdoc />
 		public DynamicUtf8PlcItem(INumericPlcItem numericPlcItem, string identifier = default)
-			: this(numericPlcItem, null, String.Empty, identifier)
+			: this(numericPlcItem, 1, null, String.Empty, identifier)
 		{ }
 
 		/// <inheritdoc />
-		public DynamicUtf8PlcItem(INumericPlcItem numericPlcItem, uint? lengthLimit, string identifier = default)
-			: this(numericPlcItem, lengthLimit, String.Empty, identifier)
+		public DynamicUtf8PlcItem(INumericPlcItem numericPlcItem, byte lengthFactor, uint? lengthLimit, string identifier = default)
+			: this(numericPlcItem, lengthFactor, lengthLimit, String.Empty, identifier)
 		{ }
 
 		/// <inheritdoc />
 		public DynamicUtf8PlcItem(INumericPlcItem numericPlcItem, string initialValue = default, string identifier = default)
-			: this(numericPlcItem, null, initialValue, identifier) { }
+			: this(numericPlcItem, 1, null, initialValue, identifier) { }
 
 		/// <inheritdoc />
-		public DynamicUtf8PlcItem(INumericPlcItem numericPlcItem, uint? lengthLimit, string initialValue = default, string identifier = default)
+		public DynamicUtf8PlcItem(INumericPlcItem numericPlcItem, byte lengthFactor, uint? lengthLimit, string initialValue = default, string identifier = default)
 			: base
 			(
 				numericPlcItem,
@@ -50,13 +50,13 @@ namespace Phoenix.Data.Plc.Items.Typed
 					new Utf8PlcItem
 					(
 						dataBlock: numericPlcItem.DataBlock,
-						position: (ushort) (numericPlcItem.Position + numericPlcItem.Value.ByteLength),
+						position: (ushort) (numericPlcItem.Position + ((IPlcItem) numericPlcItem).Value.ByteLength),
 						initialValue: initialValue,
 						canChangeSize: true,
 						identifier: name
 					),
-				lengthLimit,
-				identifier
+				lengthFactor,
+				lengthLimit, identifier
 			)
 		{ }
 
@@ -76,7 +76,7 @@ namespace Phoenix.Data.Plc.Items.Typed
 		/// <returns> A new <see cref="DynamicUtf8PlcItem"/>. </returns>
 		public new DynamicUtf8PlcItem Clone(string identifier)
 		{
-			return new DynamicUtf8PlcItem(base.LengthPlcItem, base.LengthLimit, this.Value, identifier);
+			return new DynamicUtf8PlcItem(base.LengthPlcItem, base.LengthFactor, base.LengthLimit, this.Value, identifier);
 		}
 
 		#endregion